lex_bot/requirements.txt
# Core Agent Framework
langchain
langgraph
langchain-community
langchain-core

# LLM Providers
langchain-google-genai
langchain-openai

# Vector / Embeddings
sentence-transformers
numpy

# Database
sqlalchemy
psycopg2-binary
pgvector

# Web Search & Scraping
ddgs
tavily-python
trafilatura
firecrawl-py
lxml_html_clean

# API / Server
fastapi
uvicorn
python-dotenv
requests
